Moving some common test logic into a util package.
Addresses some of #2860
