+++ kythe/go/extractors/config/BUILD
@@ -23,9 +23,9 @@ go_test(
     data = glob(["testdata/**"]),
     library = ":config",
     deps = [
+        "//kythe/go/util/testutils",
         "//kythe/proto:extraction_config_go_proto",
         "@com_github_golang_protobuf//jsonpb:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
-        "@com_github_google_go_cmp//cmp:go_default_library",
     ],
 )
